backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
pgvector>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
redis>=5.0.0
tenacity>=8.2.0
python-multipart>=0.0.6
bcrypt>=4.0.0
PyJWT>=2.8.0
httpx>=0.24.0
msgpack>=1.0.5
captcha>=0.5.0
